+++ layers/WPSchema/packages/customposts/src/Enums/CustomPostStatus.php
@@ -13,6 +13,7 @@ class CustomPostStatus
     /**
      * @todo "auto-draft" must be converted to enum value "auto_draft" on `Post.status`.
      *       Until then, this code is commented
+     * @see NonEnumerableCustomPostStatus.php
      */
     //public final const AUTO_DRAFT = 'auto_draft';
 }